LayoutTests/fast/js/sputnik/Conformance/08_Types/8.5_The_Number_Type/S8.5_A5-expected.txt

S8.5_A5

#1: NaN not greater or equal zero
#2: NaN not less or equal zero
#3: NaN not less or equal zero OR greater or equal zero
#4: NaN not less or equal zero ADD greater or equal zero
PASS 

TEST COMPLETE
